SPI: Add Cadence QSPI driver in 'direct' mode
This driver is based on an upstream driver from Altera. However, the
Altera version used the 'indirect' mode of the IP which uses optional
SRAM.

This version used the 'direct', i.e. memory mapped, mode of the IP. This
mode requires the SF layer to use memcpy to read and write data, provided
in other patches.

Signed-off-by: Michel Pollet <michel.pollet@bp.renesas.com>
Signed-off-by: Phil Edworthy <phil.edworthy@renesas.com>