mtd: nand: Driver for Evatronix NAND flash controller
This driver was sent as a patch for Linux, but since the driver APIs
are nearly identical, we can use it in U-Boot. The only change made to
the Linux patch is to not create/modify Kconfig and DT related files.
Changes required to use this driver under U-Boot are in subsequent
patches.

Currently, it supports one instance of the NANDFLASH-CTRL IP, with up
to two connected NAND flash chips, one per bank. (The IP itself is very
configurable, but the intent has been to get something working for a
specific configuration.)

This change also includes various fixes from code reviews and use on HW.

Signed-off-by: Ricard Wanderlof <ricardw@axis.com>
Signed-off-by: Phil Edworthy <phil.edworthy@renesas.com>