SPL: Make SPL use the load and entry fields from the uImage header
Note that prior to this change, the code used the entry point field
as the load address and then adjusted it by subtracting the size of
the uImage header. Why the code used the entry point instead of the
load address is anyones guess, however the offset was done to allow
the uImage to be located at the start of a NAND page and then SPL
can load whole NAND pages.

Since RZ/N1 needs to load the Cortex-M3 payload at the very start of
SRAM, the SPL ONFi code has been modified to allow unaligned offsets.

Signed-off-by: Phil Edworthy <phil.edworthy@renesas.com>