Add support for non-relocated U-Boot
U-Boot traditionally relocates itself at the top of whatever memory block
it is given, which works for large SDRAM systems, but is a waste of memory.
This is particularly an issue when using U-Boot on devices which only have
relatively small on-chip memories.

An additional benefit is that it's easier to debug U-Boot as the TEXTBASE
does not change.

This patch introduces the following symbol:
  CONFIG_SYS_STAY_IN_SRAM
It requires the board to define:
  CONFIG_SYS_SRAM_BASE
  CONFIG_SYS_SRAM_SIZE

Signed-off-by: Michel Pollet <michel.pollet@bp.renesas.com>