dfu: Handle abort correctly
Without this change, when you attempt to perform multiple downloads
and uploads to the same device/region, U-Boot will crash. The reason
for this is that U-Boot didn't reset r_left when sent an ABORT command.

Additionally, without this patch, if you perform an upload then a
download, U-Boot thinks the sequence numbers are wrong and so doesn't
write the data packet.

Signed-off-by: Phil Edworthy <phil.edworthy@renesas.com>