SF: Overhaul use of dummy cycles
The existing code calculates the number of dummy bytes based on an
incomplete list of commands, however this is completely bogus.

The SF code needs to focus on number of dummy cycles, not bytes.
Different SF devices have a different default number of dummy cycles.
These devices can usually be programmed either by volatile or
non-volatile registers to use a different number of dummy cycles, so
that operation is optimised for lower SPI clock rates.

So we allow the number of dummy cycles to be specified using the
CONFIG_SPI_FLASH_DUMMY_CYCLES symbol. Additionally, this can be overriden
using the "sf_dummy_clks" env var.

Additionally, we set the specified number of dummy cycles for Micron/ST
SPI flash devices by writing to the non-volatile register.

Signed-off-by: Phil Edworthy <phil.edworthy@renesas.com>