bootm: Provide a 2nd smp_waitloop that reads the address from a different addr
If CONFIG_SMP_PEN2_ADDR is defined, after U-Boot has kicked additional CPUs
into running _smp_pen(), U-Boot will branch to smp_waitloop2 instead of
smp_waitloop. smp_waitloop2 uses a 2nd PEN address that is used to tell the
CPU where to execute after WFI.

This is required on RZ/N1 because the BootROM gets the address to jump to
from the BOOTADDR register, but this can only be accessed in Secure mode.
Once we switch to Non-secure mode, we need a different location.

Signed-off-by: Phil Edworthy <phil.edworthy@renesas.com>