sf: Add support for memory-mapped writes for spi flashes
Some drivers allow transparently writing to a SPI flash by writing to a
memory map. This patch adds a XFER mode for MMAP_WRITE that allows the
spi slave driver to disable write protection during an explicit write,
and a separate memory map pointer for the write memory map start.

Signed-off-by: Michel Pollet <michel.pollet@bp.renesas.com>