sf: Add support for 4-byte addressing mode
If using a memory-mapped SPI Flash, you cannot use BARs to access
more than 16MB, unless you split transfers at the 16MB boundaries.

This patch switches the SPI Flash device into 4-byte addressing mode.
i.e. the commands remain the same, but the SPI Flash device expects
4-byte addresses instead of 3-bytes.
Note that this is not sensible if you do not reset the SPI Flash
device when the system is reset, since a boot ROM would normally
expect the device to use 3-byte addresses.

An alternative would be to change the commands used to the 4 byte
address versions, which appear to be common across all SF devices.
The only 4 byte address command that is not common across all
devices is Quad Page Programming. However, since the performance
of QPP is typically the same as single Page Program due to the
underlying technology of NOR flash, it should be acceptable to
use the single Page Program command.

Signed-off-by: Phil Edworthy <phil.edworthy@renesas.com>