Add support for non-relocated U-Boot
U-Boot traditionally relocates itself to the top of SDRAM as it is assumed
that U-Boot is initially run from a read-only memory, or a very small memory
that is not big enough for all the data placed on the heap.

However, some devices have a reasonable amount of on-chip SRAM, and U-Boot
is loaded into this SRAM by a BootROM. Since the size of the on-chip SRAM
is not huge, we need to save whatever memory we can, so this change allows
U-Boot to run without performing a relocation.

An additional benefit is that it's easier to debug U-Boot as the TEXTBASE
does not change.

This patch introduces the following symbol:
  CONFIG_SYS_STAY_IN_SRAM
It requires the board to define:
  CONFIG_SYS_SRAM_BASE
  CONFIG_SYS_SRAM_SIZE

Signed-off-by: Michel Pollet <michel.pollet@bp.renesas.com>
Signed-off-by: Phil Edworthy <phil.edworthy@renesas.com>