spl: rzn1: Add ability to load multiple images
This commit allows U-Boot/SPL to load multiple images from QSPI or NAND.
It uses a PKG Table loaded from an offset of CONFIG_SYS_*_U_BOOT_OFFS
from the boot media.

The PKG Table specifies whether each package is a uImage, or a Renesas
image wrapped in a BLp header. The BLp wrapped header can provide a
signature of the payload, which can be verified by the BootROM. The PKG
Table also indicates if the image contains code for the Cortex M3 core,
and if so U-Boot/SPL will start the CM3 accordingly, or data, or code
for the Cortex A7.

It contains Renesas RZ/N1 specific code as it checks that the images that
are loaded are going to a valid address, including one of the on-chip
SRAMs. Additionally, it supports the Renesas RZ/N1 proprietary SPKG image
format. U-Boot/SPL calls a function in the RZ/N1 BootROM to determine if
it must only load signed images.

Signed-off-by: Phil Edworthy <phil.edworthy@renesas.com>