arm: Make switching to HYP mode conditional
When switching to SECURE mode, instead of always switching to HYP mode
or not based on a build-time option (CONFIG_ARMV7_VIRT), introduce a
global var that can be used to specify whether we switch to HYP mode.

When running bootm, get the option to switch to HYP from a new env
variable, "boot_hyp".

Signed-off-by: Michel Pollet <michel.pollet@bp.renesas.com>