dfu: dfu_sf: Fix 4K erase
This only affects dfu with multiple targets. In this situation, the length of
data passed in may be bigger than the SF sector size. Normally, this is not
a problem because dfu_sf limits the size to the SF sector size. However, when
using multiple dfu targets, the size of the data can be determined by any one
of the dfu targets.

Signed-off-by: Phil Edworthy <phil.edworthy@renesas.com>