net: phy: marvell 883151x: Fix handling of RGMII interface types
It would appear that the 88E1518 code was only used with SGMII
interface as the code was programming the wrong registers
for tx and rx delays used with the rgmii, rgmii-txid, rgmii-rxid
and rgmii-id interface types.

Note that the PHY defaults to rgmii-id which might explain why no
one noticed before.

Tested on 88E1512 PHY which is from the same family as 88E1518.

Signed-off-by: Phil Edworthy <phil.edworthy@renesas.com>