net: dw: Handle failure in one MAC when using multiple MACs
At the moment, if there is a failure in designware_initialize(), the
Eth device is still registered. This is not a problem when you have a
single MAC, as the error will be reported back up to your board_eth_init()
function, and so the system knows there is no Ethernet available.

However, when you attempt to initialise multiple MACs, and one fails (which
is always the case when one of them tries to talk to an Extension Board that
is not present), the system will have multiple MACs registered, some of which
don't work. If it happens to be the first MAC that fails, any Ethernet cmd
will fail catastrophically when trying to talk to the PHY.

Signed-off-by: Phil Edworthy <phil.edworthy@renesas.com>