spi: cadence_qspi: Fix issue with CS toggling during memory-mapped writes
When using the memory-mapped interface, the writes do not happen with the
cadence_spi_xfer() function. Therefore, we have to ensure that the Chip
Select lines do not change at any time during other than at the start of the
transfer. Otherwise, you may find that the CS is deactivated whilst the
controller is still writing data.

Note: This is very much timing dependant as we saw no problems when the data
cache setting for the U-Boot internal data was set to write-through, but
setting it to write-back caused QSPI write errors.

Signed-off-by: Phil Edworthy <phil.edworthy@renesas.com>