rzn1: Change PTP clock from 25MHz to 125MHz
To get the best precision, use the fastest clock for PTP.
This also works around an issue that RZ/N1 Linux currently has no mechanism to
provide a clock handle for the PTP clock, and so cannot get the PTP clock rate.
The Linux gmac driver falls back to using the AXI bus clock rate for PTP which
just so happens to be 125MHz.

Signed-off-by: Phil Edworthy <phil.edworthy@renesas.com>